backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.10.6
python-dotenv==1.0.1
pytest==8.0.2
httpx==0.27.0
aiohttp==3.11.12
requests==2.31.0
pipecat-ai[deepgram,openai,cartesia,daily,silero]==0.0.101
qdrant-client==1.9.1
